% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_datos.R
\name{descargar_datos}
\alias{descargar_datos}
\title{Descargar datos meteorologicos}
\usage{
descargar_datos(id_estacion, ruta_archivo)
}
\arguments{
\item{id_estacion}{ID de la estación.}

\item{ruta_archivo}{Dirección de ruta en la que se guardara el archivo csv.}
}
\value{
Un archivo csv con los datos meteorológicos de la estación con el ID ingresado.
}
\description{
La función \code{descargar_datos} toma el ID de una estación y una dirección de ruta y descarga en dicha dirección un csv con los datos meteorologicos de la estación con dicho ID. Los IDs posibles son: NH0098, NH0046, NH437, NH472 y NH0910.
}
\examples{
descargar_datos("NH0098", "~/NH0098.csv")
descargar_datos("NH0910", "~/NH0910.csv")
}
