% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estaciones_merged}
\alias{estaciones_merged}
\title{Conjunto de datos}
\format{
Los datos son obtenidos de un repositorio de GitHub.
\describe{
\item{estaciones_merged}{muestra de datos de todas las estaciones.}
}
}
\source{
Fuente: repositorio de GitHub.
}
\usage{
estaciones_merged
}
\description{
Los datos contienen información de las distintas estaciones meteorológicas.
}
\keyword{datasets}
