% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_temperatura_mensual.R
\name{grafico_temperatura_mensual}
\alias{grafico_temperatura_mensual}
\title{Grafica de la Temperatura Mensual}
\usage{
grafico_temperatura_mensual(datos, colores = NULL, titulo = "Temperatura")
}
\arguments{
\item{datos}{data.frame a partir del cual creará los gráficos.}

\item{colores}{Colores que se utilizarán en el gráfico.}

\item{titulo}{Titulo para el gráfico.}
}
\value{
Genera un gráfico de la temperatura mensual del data.frame con el título y los colores ingresados.
}
\description{
La funcion \code{grafico_temperatura_mensual} devuelve un gráfico que muestre el promedio mensual de la temperatura de abrigo.
}
\examples{
grafico_temperatura_mensual(datos, colores, titulo)
}
