% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadatos_completos}
\alias{metadatos_completos}
\title{Septimo conjunto de datos}
\format{
Los datos son obtenidos de un repositorio de GitHub.
\describe{
\item{metadatos_completos}{muestra de metadatos de todas las estaciones}
}
}
\source{
Fuente: repositorio de GitHub.
}
\usage{
metadatos_completos
}
\description{
Los datos contienen información de las distintas estaciones meteorológicas.
}
\keyword{datasets}
