% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NH0046}
\alias{NH0046}
\title{Primer conjunto de datos}
\format{
Los datos son obtenidos de un repositorio de GitHub.
\describe{
\item{NH0046}{muestra de datos de la estación "NH0046"}
}
}
\source{
Fuente: repositorio de GitHub.
}
\usage{
NH0046
}
\description{
Los datos contienen información de las distintas estaciones meteorológicas.
}
\keyword{datasets}
