% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NH0098}
\alias{NH0098}
\title{Segundo conjunto de datos}
\format{
Los datos son obtenidos de un repositorio de GitHub.
\describe{
\item{NH0098}{muestra de datos de la estación "NH0098"}
}
}
\source{
Fuente: repositorio de GitHub.
}
\usage{
NH0098
}
\description{
Los datos contienen información de las distintas estaciones meteorológicas.
}
\keyword{datasets}
