% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NH0437}
\alias{NH0437}
\title{Tercer conjunto de datos}
\format{
Los datos son obtenidos de un repositorio de GitHub.
\describe{
\item{NH0437}{muestra de datos de la estación "NH0437"}
}
}
\source{
Fuente: repositorio de GitHub.
}
\usage{
NH0437
}
\description{
Los datos contienen información de las distintas estaciones meteorológicas.
}
\keyword{datasets}
