% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NH0472}
\alias{NH0472}
\title{Cuarto conjunto de datos}
\format{
Los datos son obtenidos de un repositorio de GitHub.
\describe{
\item{NH0472}{muestra de datos de la estación "NH0472"}
}
}
\source{
Fuente: repositorio de GitHub.
}
\usage{
NH0472
}
\description{
Los datos contienen información de las distintas estaciones meteorológicas.
}
\keyword{datasets}
