% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NH0910}
\alias{NH0910}
\title{Quinto conjunto de datos}
\format{
Los datos son obtenidos de un repositorio de GitHub.
\describe{
\item{NH0910}{muestra de datos de la estación "NH0910"}
}
}
\source{
Fuente: repositorio de GitHub.
}
\usage{
NH0910
}
\description{
Los datos contienen información de las distintas estaciones meteorológicas.
}
\keyword{datasets}
