% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_datos.R
\name{descargar_datos}
\alias{descargar_datos}
\title{Descargar datos meteorologicos}
\usage{
descargar_datos(id_estacion, directorio_destino)
}
\arguments{
\item{id_estacion}{ID de la estacion.}

\item{directorio_destino}{Directorio donde se guardara el archivo CSV.}
}
\value{
Un archivo CSV con los datos meteorologicos de la estacion con el ID ingresado.
}
\description{
La funcion \code{descargar_datos} toma el ID de una estacion y un directorio de destino, y descarga en dicha ubicacion un archivo CSV con los datos meteorologicos de la estacion correspondiente. Los IDs posibles son: NH0098, NH0046, NH437, NH472 y NH0910.
}
\examples{
descargar_datos("NH0098", "data")
descargar_datos("NH0910", "data")
}
