% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_temperatura_mensual.R
\name{grafico_temperatura_mensual}
\alias{grafico_temperatura_mensual}
\title{Gráfico de la Temperatura Mensual}
\usage{
grafico_temperatura_mensual(datos, colores = NULL, titulo = "Temperatura")
}
\arguments{
\item{datos}{data.frame a partir del cual se crearán los gráficos.}

\item{colores}{Colores que se utilizarán en el gráfico.}

\item{titulo}{Título para el gráfico.}
}
\value{
Genera un gráfico de la temperatura mensual del data.frame con el título y los colores ingresados.
}
\description{
La función \code{grafico_temperatura_mensual} devuelve un gráfico que muestra el promedio mensual de la temperatura de abrigo.
}
\examples{
# Cargar los datos de ejemplo desde el paquete
data_path <- system.file("data", "estaciones_merged.rda", package = "PaqueteMeteorologia")
load(data_path)

# Llamar a la función usando los datos cargados
grafico_temperatura_mensual(estaciones_merged, NULL, "Temperatura Mensual Promedio")
}
